requirements.txt
flask==3.0.2
flask-jwt-extended==4.6.0
prisma==0.12.0
python-dotenv==1.0.1
pandas==2.2.1
openpyxl==3.1.2
werkzeug==3.0.1 